sanctum-core/requirements.txt
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
defusedxml==0.7.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.0
fonttools==4.61.1
fpdf2==2.8.5
resend==0.6.0
greenlet>=3.1.1
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
pillow==12.0.0
psutil
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
dnspython
# NEW TESTING DEPS
pytest
httpx
pytest-cov
pyotp
qrcode
jinja2==3.1.3
playwright==1.49.1
beautifulsoup4==4.12.3